+++ src/main/java/com/booleanuk/core/products/Filling.java
@@ -1,9 +1,11 @@
 package com.booleanuk.core.products;
 
+import java.math.BigDecimal;
+
 public class Filling extends Product {
     private FillingVariant variant;
 
-    public Filling(String SKU, double price) {
+    public Filling(String SKU, BigDecimal price) {
         super(SKU, price);
     }
